+++ packages/apptive_grid_core/lib/apptive_grid_model.dart
@@ -30,6 +30,7 @@ part 'model/form/component/enum_form_component.dart';
 part 'model/form/component/enum_collection_form_component.dart';
 part 'model/form/component/cross_reference_form_component.dart';
 part 'model/form/component/geolocation_form_component.dart';
+part 'model/form/component/multi_cross_reference_form_component.dart';
 part 'model/form/component/options.dart';
 part 'model/form/form_actions.dart';
 part 'model/form/form_data.dart';